feat: parallel execution based on dataset
This commit enables parallel execution of the bin/exp-feature-sets.py
script. The script now accepts as positional arguments the name of the
dataset (eg. adult-race) & the number of iterations. The accompanying
bin/exp-feature-sets.bash script can be used to execute the python
script in parallel.

Additionally, the tests & documentation are updated to reflect the
latest changes.